fix: style & prevent refetching currencies
